pyproject.toml
[project]
name = "ward-sfx"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "einops>=0.8.1",
    "freetype-py>=2.5.1",
    "hydra-core>=1.3.2",
    "numpy>=2.3.5",
    "omegaconf>=2.3.0",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "wandb>=0.23.0",
]
